console: Create device profile container, use context, move common utils
