+++ pkg/webui/console/containers/device-onboarding-form/utils.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ACTIVATION_MODES, LORAWAN_VERSIONS, LORAWAN_PHY_VERSIONS } from '@console/lib/device-utils'
+import { ACTIVATION_MODES, hasSelectedDeviceRepositoryOther } from '@console/lib/device-utils'
 
 // End device selectors.
 
@@ -33,10 +33,6 @@ export const getJoinServerAddress = device => device.join_server_address
 
 // End device repository utils.
 
-export const SELECT_OTHER_OPTION = '_other_'
-export const isOtherOption = option => option === SELECT_OTHER_OPTION
-export const hasSelectedDeviceRepositoryOther = version =>
-  version && Object.values(version).some(value => isOtherOption(value))
 export const hasCompletedDeviceRepositorySelection = version =>
   version && Object.values(version).every(value => value)
 export const hasValidDeviceRepositoryType = (version, template) =>
@@ -58,32 +54,10 @@ export const mayProvisionDevice = (values, template) => {
   )
 }
 
-/*
-  `hardware_version` is not required when registering an end device in the device repository, so for 
-  certain end device models it can be missing. When this is the case, we still want to allow the users
-  to select such models because `firmware_version` (that might depend on hw version) and `band_id`
-  are required. `SELECT_UNKNOWN_HW_OPTION` option represents such end devices.
-*/
-export const SELECT_UNKNOWN_HW_OPTION = '_unknown_hw_version_'
-export const isUnknownHwVersion = option => option === SELECT_UNKNOWN_HW_OPTION
-
 export const REGISTRATION_TYPES = Object.freeze({
   SINGLE: 'single',
   MULTIPLE: 'multiple',
 })
-
-export const getLorawanVersionLabel = ({ lorawan_version }) => {
-  const { label } = LORAWAN_VERSIONS.find(version => version.value === lorawan_version) || {}
-
-  return label
-}
-
-export const getLorawanPhyVersionLabel = ({ lorawan_phy_version }) => {
-  const { label } =
-    LORAWAN_PHY_VERSIONS.find(version => version.value === lorawan_phy_version) || {}
-
-  return label
-}
 
 export const DEVICE_CLASS_MAP = {
   CLASS_A: 'class-a',
