+++ pkg/webui/console/containers/device-profile-section/device-card/connect.js
