+++ pkg/webui/console/containers/device-profile-section/device-card/device-card.styl
