+++ pkg/webui/console/containers/device-profile-section/device-card/index.js
