+++ pkg/webui/console/containers/device-profile-section/device-selection/brand-select/connect.js
