+++ pkg/webui/console/containers/device-profile-section/device-selection/fw-version-select/fw-version-select.js
@@ -20,12 +20,12 @@ import Select from '@ttn-lw/components/select'
 import { useFormContext } from '@ttn-lw/components/form'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
+import sharedMessages from '@ttn-lw/lib/shared-messages'
 
-import { SELECT_OTHER_OPTION, SELECT_UNKNOWN_HW_OPTION } from '../../../../utils'
-import messages from '../../../../messages'
+import { SELECT_OTHER_OPTION } from '@console/lib/device-utils'
 
 const m = defineMessages({
-  title: 'Hardware Ver.',
+  title: 'Firmware Ver.',
 })
 
 const formatOptions = (versions = []) =>
@@ -34,35 +34,26 @@ const formatOptions = (versions = []) =>
       value: version.version,
       label: version.version,
     }))
-    .concat([{ value: SELECT_OTHER_OPTION, label: messages.otherOption }])
+    .concat([{ value: SELECT_OTHER_OPTION, label: sharedMessages.otherOption }])
 
-const HardwareVersionSelect = props => {
+const FirmwareVersionSelect = props => {
   const { name, versions, onChange, ...rest } = props
   const { setFieldValue } = useFormContext()
 
-  const options = React.useMemo(() => {
-    const opts = formatOptions(versions)
-    // When only the `Other...` option is available (so end device model has no hw versions defined
-    // in the device repository) add another pseudo option that represents absence of hw versions.
-    if (opts.length === 1) {
-      opts.unshift({ value: SELECT_UNKNOWN_HW_OPTION, label: messages.unknownHwOption })
-    }
-
-    return opts
-  }, [versions])
+  const options = React.useMemo(() => formatOptions(versions), [versions])
 
   React.useEffect(() => {
     if (options.length > 0 && options.length <= 2) {
-      setFieldValue('version_ids.hardware_version', options[0].value)
+      setFieldValue('version_ids.firmware_version', options[0].value)
     }
-  }, [options, setFieldValue])
+  }, [setFieldValue, options])
 
   return (
     <Field {...rest} options={options} name={name} title={m.title} component={Select} autoFocus />
   )
 }
 
-HardwareVersionSelect.propTypes = {
+FirmwareVersionSelect.propTypes = {
   name: PropTypes.string.isRequired,
   onChange: PropTypes.func,
   versions: PropTypes.arrayOf(
@@ -72,9 +63,9 @@ HardwareVersionSelect.propTypes = {
   ),
 }
 
-HardwareVersionSelect.defaultProps = {
+FirmwareVersionSelect.defaultProps = {
   versions: [],
   onChange: () => null,
 }
 
-export default HardwareVersionSelect
+export default FirmwareVersionSelect
