+++ pkg/webui/console/containers/device-profile-section/device-selection/fw-version-select/index.js
