+++ pkg/webui/console/containers/device-profile-section/device-selection/hw-version-select/hw-version-select.js
@@ -20,12 +20,12 @@ import Select from '@ttn-lw/components/select'
 import { useFormContext } from '@ttn-lw/components/form'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
+import sharedMessages from '@ttn-lw/lib/shared-messages'
 
-import { SELECT_OTHER_OPTION } from '../../../../utils'
-import messages from '../../../../messages'
+import { SELECT_OTHER_OPTION, SELECT_UNKNOWN_HW_OPTION } from '@console/lib/device-utils'
 
 const m = defineMessages({
-  title: 'Firmware Ver.',
+  title: 'Hardware Ver.',
 })
 
 const formatOptions = (versions = []) =>
@@ -34,26 +34,35 @@ const formatOptions = (versions = []) =>
       value: version.version,
       label: version.version,
     }))
-    .concat([{ value: SELECT_OTHER_OPTION, label: messages.otherOption }])
+    .concat([{ value: SELECT_OTHER_OPTION, label: sharedMessages.otherOption }])
 
-const FirmwareVersionSelect = props => {
+const HardwareVersionSelect = props => {
   const { name, versions, onChange, ...rest } = props
   const { setFieldValue } = useFormContext()
 
-  const options = React.useMemo(() => formatOptions(versions), [versions])
+  const options = React.useMemo(() => {
+    const opts = formatOptions(versions)
+    // When only the `Other...` option is available (so end device model has no hw versions defined
+    // in the device repository) add another pseudo option that represents absence of hw versions.
+    if (opts.length === 1) {
+      opts.unshift({ value: SELECT_UNKNOWN_HW_OPTION, label: sharedMessages.unknownHwOption })
+    }
+
+    return opts
+  }, [versions])
 
   React.useEffect(() => {
     if (options.length > 0 && options.length <= 2) {
-      setFieldValue('version_ids.firmware_version', options[0].value)
+      setFieldValue('version_ids.hardware_version', options[0].value)
     }
-  }, [setFieldValue, options])
+  }, [options, setFieldValue])
 
   return (
     <Field {...rest} options={options} name={name} title={m.title} component={Select} autoFocus />
   )
 }
 
-FirmwareVersionSelect.propTypes = {
+HardwareVersionSelect.propTypes = {
   name: PropTypes.string.isRequired,
   onChange: PropTypes.func,
   versions: PropTypes.arrayOf(
@@ -63,9 +72,9 @@ FirmwareVersionSelect.propTypes = {
   ),
 }
 
-FirmwareVersionSelect.defaultProps = {
+HardwareVersionSelect.defaultProps = {
   versions: [],
   onChange: () => null,
 }
 
-export default FirmwareVersionSelect
+export default HardwareVersionSelect
