+++ pkg/webui/console/containers/device-profile-section/device-selection/hw-version-select/index.js
