+++ pkg/webui/console/containers/device-profile-section/device-selection/model-select/index.js
