+++ pkg/webui/console/containers/device-profile-section/hints/other-hint.js
