+++ pkg/webui/console/containers/device-profile-section/hints/progress-hint.js
