+++ pkg/webui/console/containers/device-profile-section/repository.styl
